Implement quest pack discovery
