+++ Assets/StreamingAssets/QuestPacks/readme.txt
@@ -1,0 +1,1 @@
+Extract quest packs to this folder before running the game.
