+++ Assets/StreamingAssets/QuestPacks/readme.txt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff12325b5090b204681dc5823d4bfe6a
+timeCreated: 1521055533
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
